+++ lib/makeOprah.js
@@ -5,7 +5,7 @@ const { log } = require('../lib/utils/logger');
 const { populateConfig, populateSecret } = require('./services');
 const { makeParameterStore } = require('./services/parameter-store/make-parameter-store');
 const { makeSettingsService } = require('./services/settings/make-settings-service');
-const { makeList } = require('./services/list-config');
+const { makeList } = require('./services/make-list');
 const { makeInit } = require('./services/make-init');
 
 const makeRun = ({ stage, interactive, getSettings, parameterStore }) => () => {
